src/python-backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
openai
numpy
